+++ server/config/index.js
@@ -5,11 +5,15 @@ export const {
   DEV_PORT,
   DB_URL,
   CLIENT_URL,
-  JWT_ACCESS_SECRET,
+  SERVER_URL,
+  ACCESS_TOKEN_SECRET,
+  ACCESS_TOKEN_EXPIRY,
   JWT_REFRESH_SECRET,
-  SERVICE,
+  REFRESH_TOKEN_SECRET,
+  REFRESH_TOKEN_EXPIRY,
+  EMAIL_SERVICE,
   EMAIL_PORT,
-  SECURE,
-  USER,
-  PASS,
+  EMAIL_SECURE,
+  EMAIL_USERID,
+  EMAIL_PASSWORD,
 } = process.env;
