--- server/controllers/index.js
@@ -1,10 +1,0 @@
-export { default as signupController } from "./signupController";
-export { default as loginController } from "./loginController";
-export { default as refreshController } from "./refreshController";
-export { default as userController } from "./userController";
-export { default as logoutController } from "./logoutController";
-export { default as adminController } from "./admin/adminController";
-export { default as shopController } from "./shopController";
-export { default as bookDetailsController } from "./bookDetailsController";
-export { default as orderController } from "./orderController";
-export { default as contactUsController } from "./contactUsController";
