+++ server/env-template.txt
@@ -1,15 +1,15 @@
 DEV_PORT=DEV_PORT
 DB_URL=MONGODB_URL
-JWT_ACCESS_SECRET=JWT_ACCESS_SECRET
-JWT_REFRESH_SECRET=JWT_REFRESH_SECRET
+ACCESS_TOKEN_SECRET=ACCESS_TOKEN_SECRET
+REFRESH_TOKEN_SECRET=REFRESH_TOKEN_SECRET
 CLIENT_URL=FRONTEND_URL
 
 # FOR SENDING EMAIL FROM NODEMAILER
 
 # REFER https://www.youtube.com/watch?v=lBRnLXwjLw0
 
-SERVICE=SERVICE
+EMAIL_SERVICE=SERVICE
 EMAIL_PORT=PORT_NUMEBR
 SECURE=SECURE
-USER=YOUR_EMAIL_ID
-PASS=16_DEGIT_CODE_FROM_YOUR_GMAIL_ACCOUNT
+EMAIL_USERID=YOUR_EMAIL_ID
+EMAIL_PASSWORD=16_DEGIT_CODE_FROM_YOUR_GMAIL_ACCOUNT
