+++ server/index.js
@@ -1,14 +1,14 @@
 import express from "express";
-import cors from "cors";
 import { DEV_PORT, DB_URL, CLIENT_URL } from "./config";
 import { errorHandler } from "./middlewares";
 import routes from "./routes";
 import mongoose from "mongoose";
-import cookieParser from "cookie-parser";
+// import cors from "cors";
+// import cookieParser from "cookie-parser";
 
 const app = express();
-app.use(cors({ origin: CLIENT_URL, credentials: true }));
-app.use(cookieParser());
+// app.use(cors({ origin: CLIENT_URL, credentials: true }));
+// app.use(cookieParser());
 
 // Database connection
