+++ server/middlewares/auth.js
@@ -1,24 +1,25 @@
-import { CustomErrorHandler, JwtService } from "../services";
-
+import jwt from "jsonwebtoken";
+import { ACCESS_TOKEN_SECRET } from "../config";
+import { customErrorHandler } from "../services";
 const auth = async (req, res, next) => {
   // Accessing the access token from headers
   let authHeader = req.headers.authorization;
   if (!authHeader) {
-    return next(CustomErrorHandler.unAuthorized("accessToken Required"));
+    return next(customErrorHandler.unAuthorized("accessToken Required"));
   }
   const authAccessToken = authHeader.split(" ")[1];
 
   //Verifying the access token
   try {
-    const { _id, role } = JwtService.verify(authAccessToken);
+    const { userId, role } = jwt.verify(ACCESS_TOKEN_SECRET);
     const user = {
-      _id,
+      userId,
       role,
     };
     req.user = user;
     next();
   } catch (err) {
-    return next(CustomErrorHandler.unAuthorized(err.message));
+    return next(customErrorHandler.unAuthorized(err.message));
   }
 };
