+++ server/middlewares/errorHandler.js
@@ -1,5 +1,5 @@
 import { ValidationError } from "joi";
-import { CustomErrorHandler } from "../services";
+import { customErrorHandler } from "../services";
 
 const errorHandler = (err, req, res, next) => {
   let statusCode = 500;
@@ -13,7 +13,7 @@ const errorHandler = (err, req, res, next) => {
       message: err.message,
     };
   }
-  if (err instanceof CustomErrorHandler) {
+  if (err instanceof customErrorHandler) {
     statusCode = err.status;
     data = {
       message: err.message,
