--- server/models/index.js
@@ -1,5 +1,0 @@
-export { default as User } from "./user";
-export { default as RefreshToken } from "./refreshToken";
-export { default as Book } from "./book";
-export { default as Order } from "./order";
-export { default as EmailToken } from "./emailToken";
