+++ server/models/user.js
@@ -12,4 +12,16 @@ const userSchema = new mongoose.Schema(
   { timestamp: true }
 );
 
-export default mongoose.model("User", userSchema, "users");
+const virtual = userSchema.virtual("id");
+virtual.get(function () {
+  return this._id;
+});
+userSchema.set("toJSON", {
+  virtuals: true,
+  versionKey: false,
+  transform: function (doc, ret) {
+    delete ret._id;
+  },
+});
+
+export const User = mongoose.model("User", userSchema, "users");
