+++ server/routes/index.js
@@ -1,47 +1,43 @@
 import express from "express";
 import { auth, admin } from "../middlewares";
+import { signup } from "../controllers/auth/signup";
+import { login } from "../controllers/auth/login";
+import { verifyEmail, sendEmail } from "../controllers/auth/verifyEmail";
+import { changePassword } from "../controllers/auth/changePassword";
+import { tokenRefresh } from "../controllers/auth/tokenRefresh";
+import { logout } from "../controllers/auth/logout";
+import { contactUs } from "../controllers/contactUs";
+import { getUserById, editUser, deleteUser } from "../controllers/user";
+import { getAllOrders, placeOrder } from "../controllers/order";
 import {
-  signupController,
-  loginController,
-  refreshController,
-  userController,
-  logoutController,
-  adminController,
-  shopController,
-  bookDetailsController,
-  orderController,
-  contactUsController,
-} from "../controllers";
+  getBookById,
+  getAllBooks,
+  getBestSellingBooks,
+} from "../controllers/book";
+import { addBook, deleteBook } from "../controllers/admin";
 
 const router = express.Router();
 
-router.post(
-  "/signup",
-  signupController.signup,
-  signupController.sendVerificationEmail
-);
-router.post("/sendverificationemail", signupController.sendVerificationEmail);
-router.get(
-  "/verifyemail/:emailverificationtoken",
-  signupController.verifyEmail
-);
-router.post("/login", loginController.login);
-router.post("/changepassword", auth, loginController.changePassword);
-router.post("/logout", logoutController.logout);
-router.post("/refresh", refreshController.refresh);
+router.post("/signup", signup, sendEmail);
+router.get("/verify/:emailtoken", verifyEmail);
+router.post("/login", login);
+router.post("/forgotpassword", sendEmail);
+router.post("/changepassword", auth, changePassword);
+router.post("/refresh", tokenRefresh);
+router.post("/logout", logout);
+router.post("/contactus", contactUs);
 
-router.get("/user", auth, userController.user);
-router.post("/updateemail", auth, userController.updateEmail);
-router.get("/deleteuser", auth, userController.deleteUser);
-router.get("/getorders", auth, orderController.getOrders);
+router.get("/user", auth, getUserById);
+router.post("/updateemail", auth, editUser);
+router.get("/deleteuser", auth, deleteUser);
 
-router.get("/bestsellingbooks", shopController.bestSellingBooks);
-router.post("/shop/:currentPage", shopController.shop);
-router.get("/bookdetails/:bookid", bookDetailsController.getbook);
-router.post("/placeorder", auth, orderController.placeOrder);
-router.post("/contactus", contactUsController.contactUs);
+router.get("/bookdetails/:bookid", getBookById);
+router.get("/bestsellingbooks", getBestSellingBooks);
+router.post("/shop/:currentPage", getAllBooks);
+router.get("/getorders", auth, getAllOrders);
+router.post("/placeorder", auth, placeOrder);
 
-router.post("/admin", [auth, admin], adminController.addBook);
-router.delete("/deletebook/:bookid", [auth, admin], adminController.deleteBook);
+router.post("/admin", [auth, admin], addBook);
+router.delete("/deletebook/:bookid", [auth, admin], deleteBook);
 
 export default router;
